Use naming service directly to find event channel in ECM tests.
Tests were using a "corbaname" URL to check the existence of an event channel,
which implicitly depended on the CORBA name service being on localhost. Since
the name service is already available to the test, it's more robust to resolve
names that way, which also allows the use of a GitLab service for omniNames in
CI jobs.